pyproject.toml
[project]
name = "callmind"
version = "0.1.0"
description = "AI Sales Memory That Learns — upload call recordings, extract insights, Q-learning ranks what matters"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Ivan Pasichnyk", email = "ivan@welabeldata.com"},
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.12",
    "google-genai>=1.0.0",
    "yt-dlp>=2024.12.0",
    "httpx>=0.27.0",
    "qdrant-client>=1.12.0",
    "fastembed>=0.4.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[project.scripts]
callmind = "callmind.app:main"
